back_end/Library_django/requirement.txt
Django==1.10
django-ckeditor==5.3.0
django-cors-headers==2.1.0
django-filter==1.0.4
django-js-asset==0.1.1
djangorestframework==3.6.4
Markdown==2.6.9
olefile==0.44
Pillow==4.2.1
